requirements.txt
opencv-python==4.8.0.74
matplotlib==3.7.2
numpy
Pillow
